🎨 useGetMe 커스텀 훅 생성하여 기존 getMe 함수 대체
